refactor(architecture): enhance folder structure for course project to mirror component imports
