+++ apps/course/src/app/app.component.html
@@ -7,7 +7,7 @@
   <router-outlet />
 </div>
 @if (toastService.getToasts().length) {
-  @for (toast of toastService.getToasts();track $index) {
+  @for (toast of toastService.getToasts(); track $index) {
     <div class="toast">
       <div
         class="alert text-base-100"
