+++ apps/course/src/app/features/home/courses/courses.component.html
@@ -4,7 +4,7 @@
     <app-loading-bars />
   } @else if (!isLoading && courses.length) {
     <div class="flex flex-col">
-      @for (course of courses;track $index) {
+      @for (course of courses; track $index) {
         <div class="my-1.5"></div>
         <div
           role="link"
