+++ apps/course/src/app/features/home/courses/courses.component.ts
@@ -2,9 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router, RouterLink } from '@angular/router';
 import { CourseService } from '../../../core/data/course.service';
 import { Course } from '../../../../../projects/types/src/lib/course.types';
-import { LoadingBarsComponent } from '../../../shared/loading-bars/loading-bars.component';
+import { LoadingBarsComponent } from '../../../shared/feedback/loading-bars/loading-bars.component';
 import { CoursesFilterComponent } from './filter/courses-filter.component';
-import { FeedbackMessageComponent } from '../../../shared/feedback-message/feedback-message.component';
+import { FeedbackMessageComponent } from '../../../shared/feedback/feedback-message/feedback-message.component';
 
 @Component({
   selector: 'app-courses',
