+++ apps/course/src/app/shared/feedback/loading-bars/loading-bars.component.ts
