+++ apps/course/src/app/shared/navigation/footer/footer.component.ts
