+++ apps/course/src/app/shared/navigation/navbar-course/navbar-course.component.html
