+++ apps/course/src/app/shared/navigation/navbar-course/navbar-course.component.spec.ts
