+++ apps/course/src/app/shared/navigation/navbar-course/navbar-course.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { environment } from '../../../environments/environment';
+import { environment } from '../../../../environments/environment';
 import { RouterLink } from '@angular/router';
 
 @Component({
