+++ apps/course/src/app/shared/navigation/navbar/navbar.component.html
